requirements.txt
langchain-google-genai
langchain-mcp-adapters
mcp
langgraph
asyncpg
python-dotenv